requirements.txt
mcp>=1.3.0
msgraph-sdk>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
msal>=1.22.0
uvicorn>=0.23.0
pydantic>=2.0.0 